Heatmap Calculation error with equal appVersion name
If the program starts with a emtpy project and then import a knxprod (e.g. MDT_KP_BE_01_Universal_Interface_V16a.knxprod) there are many application programs with equal AppVersion Names (all are "Imported")
In this case , the heatmap can't be calculated (errors are shown)
Therefore the Names of the AppVersions have to be differently.
So, my suggestion is to set the AppVerion name to the name of the application program with "_Imported" as attachment.